feat(zoom): relocate zoom-hooks
