+++ lib/Components/Zoom/index.ts
@@ -1,11 +1,8 @@
 export type {
-    ZoomOptions,
     ZoomInButtonProps,
     ZoomOutButtonProps,
     ZoomFitButtonProps,
 } from "./Zoom.types";
-
-export { default as useZoom } from "./useZoom";
 
 export { default as ZoomInButton } from "./ZoomInButton";
 export { default as ZoomOutButton } from "./ZoomOutButton";
