+++ lib/Control/Zoom/ZoomFitControl.tsx
@@ -2,7 +2,7 @@ import React from "react";
 
 
 import type { ZoomControlProps } from "./Zoom.types";
-import useAttachedZoom from "./useAttachedZoom";
+import { useAttachedZoom } from "../../Diagram";
 import { ZoomFitButton } from "../../Components/Zoom";
 
 const ZoomFitControl = ({ diagram, options, title }: ZoomControlProps) => {
