+++ lib/Control/Zoom/index.ts
@@ -1,12 +1,9 @@
 export type {
-    AttachedZoomOptions,
     ZoomControlProps,
     ZoomControlGroupProps,
 } from "./Zoom.types";
 
-export { default as useAttachedZoom } from "./useAttachedZoom";
-
 export { default as ZoomInControl } from "./ZoomInControl";
 export { default as ZoomOutControl } from "./ZoomOutControl";
-export { default as ZoomResetControl } from "./ZoomFitControl";
+export { default as ZoomFitControl } from "./ZoomFitControl";
 export { default as ZoomControlGroup } from "./ZoomControlGroup";
