+++ lib/Diagram/hooks/index.ts
@@ -1,7 +1,9 @@
 export type {
+    AttachedZoomOptions,
     EventBusEventCallback,
     Overlay,
 } from "./hooks.types";
 
+export { default as useAttachedZoom } from "./useAttachedZoom";
 export { default as useEventHandler } from "./useEventHandler";
 export { default as useOverlays } from "./useOverlays";
