+++ stories/Diagram/useAttachedZoom.mdx
@@ -1,8 +1,3 @@
-import { Meta } from "@storybook/blocks";
-import LinkTo from '@storybook/addon-links/react';
-
-<Meta title="Controls/Zoom/useAttachedZoom" />
-
 # Overview
 
 This hook handles the integration with the zoom API of `diagram-js` and allows modifying the current zoom while staying in sync with external changes to the zoom level.
